Add platform::IsMemoryValid.
